pyproject.toml
[project]
name = "my-info-vw"
version = "0.1.0"
description = "AI-powered message fact-checker using multi-channel verification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.12",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
